Adjustements after review
