+++ ReactAndroid/src/main/java/com/facebook/react/uimanager/layoutanimation/LayoutAnimationListener.java
@@ -1,0 +1,10 @@
+// Copyright 2004-present Facebook. All Rights Reserved.
+
+package com.facebook.react.uimanager.layoutanimation;
+
+/**
+ * Listener invoked when a layout animation has completed.
+ */
+public interface LayoutAnimationListener {
+  void onAnimationEnd();
+}
